+++ tests/unit/controllers/CartTest.php
@@ -9,16 +9,23 @@ namespace craftcommercetests\unit\controllers;
 
 use Codeception\Test\Unit;
 use Craft;
+use craft\commerce\behaviors\CustomerBehavior;
 use craft\commerce\controllers\CartController;
+use craft\commerce\elements\Product;
 use craft\commerce\elements\Variant;
 use craft\commerce\Plugin;
+use craft\elements\User;
 use craft\errors\ElementNotFoundException;
 use craft\errors\InvalidPluginException;
 use craft\web\Request;
+use craftcommercetests\fixtures\CustomerAddressFixture;
+use craftcommercetests\fixtures\CustomerFixture;
+use craftcommercetests\fixtures\ProductFixture;
 use craftcommercetests\fixtures\SalesFixture;
 use Throwable;
 use UnitTester;
 use yii\base\Exception;
+use yii\base\InvalidConfigException;
 use yii\base\InvalidRouteException;
 use yii\web\Response;
 
@@ -51,8 +58,17 @@ class CartTest extends Unit
     public function _fixtures(): array
     {
         return [
+            'products' => [
+                'class' => ProductFixture::class,
+            ],
             'sales' => [
                 'class' => SalesFixture::class,
+            ],
+            'customer' => [
+                'class' => CustomerFixture::class,
+            ],
+            'addresses' => [
+                'class' => CustomerAddressFixture::class,
             ],
         ];
     }
@@ -276,7 +292,7 @@ class CartTest extends Unit
      * @throws InvalidRouteException
      * @throws Throwable
      * @throws \craft\errors\InvalidFieldException
-     * @throws \yii\base\InvalidConfigException
+     * @throws InvalidConfigException
      */
     public function testAddAddressCustomFieldsOnUpdateCart(): void
     {
@@ -308,5 +324,65 @@ class CartTest extends Unit
         self::assertEquals($shippingAddress['fields']['testPhone'], $cartShippingAddress->testPhone);
         self::assertEquals($billingAddress['addressLine1'], $cartBillingAddress->addressLine1);
         self::assertEquals($billingAddress['fields']['testPhone'], $cartBillingAddress->testPhone);
+    }
+
+    /**
+     * @return void
+     * @throws ElementNotFoundException
+     * @throws Exception
+     * @throws InvalidPluginException
+     * @throws InvalidRouteException
+     * @throws Throwable
+     * @throws InvalidConfigException
+     * @since 4.0.4
+     */
+    public function testAutoSetNewCartAddresses(): void
+    {
+        Craft::$app->getPlugins()->switchEdition('commerce', Plugin::EDITION_PRO);
+        $this->request->headers->set('X-Http-Method-Override', 'POST');
+
+        $customerFixture = $this->tester->grabFixture('customer');
+        /** @var User|CustomerBehavior $customer */
+        $customer = $customerFixture->getElement('customer3');
+        Craft::$app->getUser()->setIdentity(
+            Craft::$app->getUsers()->getUserById($customer->id)
+        );
+        $customerShippingAddress = $customer->getPrimaryShippingAddress();
+
+        $productsFixture = $this->tester->grabFixture('products');
+        /** @var Product $product */
+        $product = $productsFixture->getElement('rad-hoodie');
+        $bodyParams = [
+            'purchasableId' => $product->getDefaultVariant()->id,
+            'qty' => 2,
+        ];
+        $this->request->setBodyParams($bodyParams);
+
+        $this->cartController->runAction('update-cart');
+
+        $cart = Plugin::getInstance()->getCarts()->getCart();
+        $shippingAddress = $cart->getShippingAddress();
+
+        self::assertEquals($customerShippingAddress->addressLine1, $shippingAddress->addressLine1);
+
+        Plugin::getInstance()->getCarts()->forgetCart();
+        Craft::$app->getElements()->deleteElement($cart, true);
+
+        $this->request->setBodyParams($bodyParams);
+
+        $originalSettingValue = Plugin::getInstance()->getSettings()->autoSetNewCartAddresses;
+
+        Plugin::getInstance()->getSettings()->autoSetNewCartAddresses = false;
+
+        $this->cartController->runAction('update-cart');
+
+        $cart = Plugin::getInstance()->getCarts()->getCart();
+        $shippingAddress = $cart->getShippingAddress();
+
+        self::assertNull($shippingAddress);
+        Plugin::getInstance()->getCarts()->forgetCart();
+        Craft::$app->getElements()->deleteElement($cart, true);
+
+        Plugin::getInstance()->getSettings()->autoSetNewCartAddresses = $originalSettingValue;
     }
 }