+++ tests/fixtures/data/customer-addresses.php
@@ -1,0 +1,17 @@
+<?php
+
+return [
+    'box' => [
+        '_owner' => 'customer3',
+        '_primary' => ['shipping'],
+        'firstName' => 'John',
+        'lastName' => 'Doe',
+        'addressLine1' => '23 Woodworth ',
+        'addressLine2' => 'Seaside',
+        'locality' => 'County Island',
+        'postalCode' => '12345',
+        'title' => 'TV Show',
+        'countryCode' => 'US',
+        'administrativeArea' => 'NY',
+    ],
+];